# openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "GEOS API",
  "version": "0.3.0",
  "description": "Куда пойти именно тебе: рекомендации мест по вкусу с причинами, настоящее состояние места, лиды."
 },
 "servers": [
  {
   "url": "https://geos-six.vercel.app"
  }
 ],
 "paths": {
  "/api/recommend": {
   "get": {
    "operationId": "recommend",
    "summary": "Ранжированные места рядом с точкой по вкусу",
    "parameters": [
     {
      "name": "lat",
      "in": "query",
      "schema": {
       "type": "number"
      }
     },
     {
      "name": "lon",
      "in": "query",
      "schema": {
       "type": "number"
      }
     },
     {
      "name": "n",
      "in": "query",
      "schema": {
       "type": "integer"
      }
     },
     {
      "name": "cat",
      "in": "query",
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "taste",
      "in": "query",
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "radius",
      "in": "query",
      "schema": {
       "type": "number"
      }
     },
     {
      "name": "project",
      "in": "query",
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "sponsored_only",
      "in": "query",
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "items[] с id, name, cat, lat, lon, dist_km, score, why[], sponsored, open, phone, site, hours"
     }
    }
   }
  },
...
 }
}